refactoring cockpit
