+++ app/src/App.svelte
@@ -3,7 +3,7 @@
   import { Router, Route } from "svelte-navigator";
   import { getMarketAndIDL } from './scripts/jet';
   import { getLocale } from './scripts/localization';
-  import { initDarkTheme } from './scripts/util';
+  import { setDark } from './scripts/util';
   import Nav from './components/Nav.svelte';
   import Cockpit from './views/Cockpit.svelte';
   import TransactionLogs from "./views/TransactionLogs.svelte";
@@ -13,10 +13,11 @@
   import Copilot from './components/Copilot.svelte';
   import Notifications from './components/Notifications.svelte';
 
+  let darkTheme: boolean = localStorage.getItem('jetDark') === 'true';
   let launchUI: boolean = false;
   onMount(async () => {
     // Initialize dark theme
-    initDarkTheme();
+    if (darkTheme) setDark(true);
     // Get user's locale and check for banned region
     await getLocale();
     // get IDL whith market reserve data
