+++ app/src/components/Logo.svelte
@@ -3,12 +3,12 @@
   export let logoMark: boolean = false;
 </script>
 
-<div class="logo flex align-center justify-center">
+<div class="logo flex-centered">
   <img src="img/jet/jet_logomark_gradient.png" alt="Jet Logomark"
-    style={`width: ${width}px; ${logoMark ? 'height: auto; opacity: 1' : 'height: 0px; opacity: 0; position: absolute;'}`} 
+    style="width: {width}px; {logoMark ? 'height: auto; opacity: 1' : 'height: 0px; opacity: 0; position: absolute;'}" 
   />
   <img src="img/jet/jet_logo_gradient.png" alt="Jext Logo"
-    style={`width: ${width}px; ${!logoMark ? 'height: auto; opacity: 1' : 'height: 0px; opacity: 0; position: absolute;'}`} 
+    style="width: {width}px; {!logoMark ? 'height: auto; opacity: 1' : 'height: 0px; opacity: 0; position: absolute;'}" 
   />
 </div>
