+++ app/src/components/Notifications.svelte
@@ -6,13 +6,13 @@
 </script>
 
 {#if $USER.notifications?.length}
-  <div class="notifications flex align-center justify-center column">
+  <div class="notifications flex-centered column">
     {#each $USER.notifications as n, i}
-      <div class="notification flex align-center justify-center"
+      <div class="notification flex-centered"
         class:success={n.success}
         in:fly={{y: 50, duration: 500}}
         out:fade={{duration: 50}}>
-        <div class="copilot-img flex align-center justify-center"
+        <div class="copilot-img flex-centered"
           on:click={() => {if (n.success) navigate("/transactions")}}>
           <img src="img/copilot/copilot.png" 
             alt="Copilot Icon"
